requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
itsdangerous==2.2.0
python-multipart==0.0.18
fastapi-mail==1.4.2
slowapi==0.1.9
jinja2==3.1.4
alembic==1.17.2